frag to activities

+++ app/src/main/java/com/example/medicaldiagnosisapp/DiagnoseActivityPage1.java
@@ -19,7 +19,7 @@ public class DiagnoseActivityPage1 extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_diagnose_page1);
 
-        final int [] diagnoseArr = {0,0,0,0,0,0,0,0,0,0,0,0};
+        final double [] diagnoseArr = {0,0,0,0,0,0,0,0,0,0,0,0};
         //array is [phy,ingest,allergy,heart,stroke,poison,heat,anaphy,burn,trauma,bone]
